Add provider to host
