+++ packages/common/src/api/OmSupplyApiContext.tsx
@@ -1,7 +1,17 @@
 import React, { FC, useEffect, useState, useCallback } from 'react';
 import { createContext } from 'react';
-import { OmSupplyApi, createOmSupplyApi } from '.';
 import { GraphQLClient } from 'graphql-request';
+import { getSdk } from '../types';
+
+export const createOmSupplyApi = (
+  url: string
+): { api: OmSupplyApi; client: GraphQLClient } => {
+  const client = new GraphQLClient(url);
+  const api = getSdk(client);
+  return { client, api };
+};
+
+export type OmSupplyApi = ReturnType<typeof getSdk>;
 
 interface OmSupplyApiControl {
   api: OmSupplyApi;
